🐛 Fixed reloading paths (#17)
There was a problem with checking routes.

If `currentPath` and `localePath` in `common.astro` layout are different, it redirects user to the locale one. The problem was in a situation, when the localized path or the current path have or have not a trailing slash - this resulted in constant reloading of the page because of a difference like `SITE/page` vs `SITE/page/`.

Added some conditional logic in there to prevent those situations.